Separate auto-wiring tests into another file
